% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_monochrome}
\alias{get_monochrome}
\title{Get Monochomatic Colors}
\usage{
get_monochrome()
}
\value{
A vector of monochromatic colors
}
\description{
This function loads the monochromatic colors in our style guide.
}
