% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_sidebyside}
\alias{get_sidebyside}
\title{Get Side by Side Colors}
\usage{
get_sidebyside()
}
\value{
A vector of Side by Side colors
}
\description{
This function loads the Side by Side colors in our style guide.
}
